fix using the system language if no language is selected also move the core string from ipc into deltachat/login
