+++ src/main/deltachat/login.ts
@@ -1,4 +1,4 @@
-import { DeltaChat } from 'deltachat-node'
+import { DeltaChat, C } from 'deltachat-node'
 import { app as rawApp } from 'electron'
 import { getLogger } from '../../shared/logger'
 import { setupMarkseenFix } from '../markseenFix'
@@ -6,7 +6,6 @@ import setupNotifications from '../notifications'
 import setupUnreadBadgeCounter from '../unread-badge'
 import SplitOut from './splitout'
 import { Credentials, DeltaChatAccount } from '../../shared/shared-types'
-import { txCoreStrings } from '../ipc'
 import { getNewAccountPath, getLogins, removeAccount } from '../logins'
 import { ExtendedAppMainProcess } from '../types'
 import { serverFlags } from './settings'
@@ -205,4 +204,47 @@ Full changelog: https://github.com/deltachat/deltachat-desktop/blob/master/CHANG
     }
     return selectedAccount
   }
+}
+
+export function txCoreStrings() {
+  const tx = app.translate
+  const strings: { [key: number]: string } = {}
+  // TODO: Check if we need the uncommented core translations
+  strings[C.DC_STR_NOMESSAGES] = tx('chat_no_messages')
+  strings[C.DC_STR_SELF] = tx('self')
+  strings[C.DC_STR_DRAFT] = tx('draft')
+  strings[C.DC_STR_VOICEMESSAGE] = tx('voice_message')
+  strings[C.DC_STR_DEADDROP] = tx('login_inbox')
+  strings[C.DC_STR_IMAGE] = tx('image')
+  strings[C.DC_STR_GIF] = tx('gif')
+  strings[C.DC_STR_VIDEO] = tx('video')
+  strings[C.DC_STR_AUDIO] = tx('audio')
+  strings[C.DC_STR_FILE] = tx('file')
+  strings[C.DC_STR_ENCRYPTEDMSG] = tx('encrypted_message')
+  strings[C.DC_STR_STATUSLINE] = tx('pref_default_status_text')
+  strings[C.DC_STR_NEWGROUPDRAFT] = tx('group_hello_draft')
+  strings[C.DC_STR_MSGGRPNAME] = tx('systemmsg_group_name_changed')
+  strings[C.DC_STR_MSGGRPIMGCHANGED] = tx('systemmsg_group_image_changed')
+  strings[C.DC_STR_MSGADDMEMBER] = tx('systemmsg_member_added')
+  strings[C.DC_STR_MSGDELMEMBER] = tx('systemmsg_member_removed')
+  strings[C.DC_STR_MSGGROUPLEFT] = tx('systemmsg_group_left')
+  // strings[C.DC_STR_E2E_AVAILABLE] = tx('DC_STR_E2E_AVAILABLE')
+  // strings[C.DC_STR_ENCR_TRANSP] = tx('DC_STR_ENCR_TRANSP')
+  // strings[C.DC_STR_ENCR_NONE] = tx('DC_STR_ENCR_NONE')
+  strings[C.DC_STR_FINGERPRINTS] = tx('qrscan_fingerprint_label')
+  strings[C.DC_STR_READRCPT] = tx('systemmsg_read_receipt_subject')
+  strings[C.DC_STR_READRCPT_MAILBODY] = tx('systemmsg_read_receipt_body')
+  strings[C.DC_STR_MSGGRPIMGDELETED] = tx('systemmsg_group_image_deleted')
+  strings[C.DC_STR_E2E_PREFERRED] = tx('autocrypt_prefer_e2ee')
+  strings[C.DC_STR_ARCHIVEDCHATS] = tx('chat_archived_chats_title')
+  // strings[C.DC_STR_STARREDMSGS] = tx('DC_STR_STARREDMSGS')
+  strings[C.DC_STR_AC_SETUP_MSG_SUBJECT] = tx('autocrypt_asm_subject')
+  strings[C.DC_STR_AC_SETUP_MSG_BODY] = tx('autocrypt_asm_general_body')
+  strings[C.DC_STR_CANTDECRYPT_MSG_BODY] = tx('systemmsg_cannot_decrypt')
+  strings[C.DC_STR_CANNOT_LOGIN] = tx('login_error_cannot_login')
+  strings[C.DC_STR_SERVER_RESPONSE] = tx('login_error_server_response')
+  strings[68] = tx('device_talk')
+  strings[69] = tx('saved_messages')
+
+  return strings
 }
